// src/types/requestCriterias.ts
export enum MedicationTypeLabel {
  Request = 'Prescription',
  Administration = 'Administration'
}

export type LabelCriteriaObject = {
  id: string
  label: string
  system?: ValueSetSystem
}

export enum RessourceType {
  REQUEST = 'Request',
  IPP_LIST = 'IPPList',
  PATIENT = 'Patient',
  ENCOUNTER = 'Encounter',
  DOCUMENTS = 'DocumentReference',
  PMSI = 'pmsi',
  CONDITION = 'Condition',
  PROCEDURE = 'Procedure',
  CLAIM = 'Claim',
  MEDICATION = 'Medication',
  MEDICATION_REQUEST = 'MedicationRequest',
  MEDICATION_ADMINISTRATION = 'MedicationAdministration',
  BIO_MICRO = 'biologie_microbiologie',
  OBSERVATION = 'Observation',
  MICROBIOLOGIE = 'microbiologie',
  PHYSIOLOGIE = 'physiologie',
  IMAGING = 'ImagingStudy'
}

export enum RessourceTypeLabels {
  REQUEST = 'Mes requềtes',
  IPP_LIST = "Liste d'IPP",
  PATIENT = 'Démographie',
  ENCOUNTER = 'Prise en charge',
  DOCUMENTS = 'Documents cliniques',
  PMSI = 'PMSI',
  CONDITION = 'Diagnostic',
  PROCEDURE = 'actes CCAM',
  CLAIM = 'GHM',
  MEDICATION = 'Médicament',
  BIO_MICRO = 'Biologie/Microbiologie',
  OBSERVATION = 'Biologie',
  MICROBIOLOGIE = 'Microbiologie',
  PHYSIOLOGIE = 'Physiologie',
  IMAGING = 'Imagerie'
}

export type CommonCriteriaDataType = {
  id: number
  error?: boolean
  type: RessourceType
  encounterService?: ScopeTreeRow[]
  isInclusive?: boolean
  title: string
}
// ...
export enum Comparators {
  LESS_OR_EQUAL = '<=',
  LESS = '<',
  EQUAL = '=',
  GREATER = '>',
  GREATER_OR_EQUAL = '>=',
  BETWEEN = '<x>'
}
// ...
export type MedicationDataType = CommonCriteriaDataType & {
  code: LabelObject[] | null
  prescriptionType: LabelObject[] | null
  administration: LabelObject[] | null
  occurrence: number
  occurrenceComparator: Comparators
  startOccurrence: string | null
  endOccurrence: string | null
  encounterEndDate: string | null
  encounterStartDate: string | null
} & (
    | {
        type: RessourceType.MEDICATION_REQUEST
        prescriptionType: LabelObject[] | null
      }
    | { type: RessourceType.MEDICATION_ADMINISTRATION }
  )